pyproject.toml
[tool.poetry]
name = "email-sender2"
version = "0.1.0"
description = ""
authors = ["DaniilMashkov <dankeeeez@gmail.com>"]
readme = "README.md"
packages = [{include = "email_sender2"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
django-crispy-forms = "^1.14.0"
django-crontab = "^0.7.1"
django-redis = "^5.2.0"
pillow = "^9.4.0"
python-dotenv = "^0.21.1"
django-hitcount = "^1.3.5"
django-utils-six = "^2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
